return value of strlen casted to int now.

+++ src/global/usage.c
@@ -1,6 +1,9 @@
 /*
  *
  * $Log$
+ * Revision 3.44  2003/04/14 13:51:05  sbs
+ * return value of strlen casted to int now.
+ *
  * Revision 3.43  2003/03/24 16:36:52  sbs
  * cppI added
  *
@@ -132,7 +135,7 @@
     {                                                                                    \
         int _i;                                                                          \
         printf ("    -b %2i:%s", ph, spec);                                              \
-        for (_i = 0; _i < (12 - strlen (spec)); _i++) {                                  \
+        for (_i = 0; _i < (12 - (int)strlen (spec)); _i++) {                             \
             printf (" ");                                                                \
         }                                                                                \
         printf ("%s\n", comment);                                                        \
